remove unnecessary casts in "935-ver1"

+++ algorithm/688-Knight-Probability-in-Chessboard/ver1.cpp
@@ -39,6 +39,6 @@ public:
             }
         }
         
-        return static_cast<double>(totSteps) / pow(8, K);
+        return totSteps / pow(8, K);
     }
 };
